Update to 151.7896.
Update to commit 6dd4e9534d434ef9dad5054178912af38d34c682

+++ chrome/browser/ui/webui/new_tab_page/new_tab_page_handler_browsertest.cc
@@ -232,8 +232,8 @@ IN_PROC_BROWSER_TEST_F(
 IN_PROC_BROWSER_TEST_F(
     NewTabPageHandlerWithCustomizeChromePromoMaxTimesBrowserTest,
     DISABLED_DontOpenPanelWhenPanelWasShowedMaxTimesBefore) {
-  for (size_t i = 0;
-       i < ntp_features::kNtpCustomizeChromeAutoShownMaxCount.Get(); ++i) {
+  for (int i = 0; i < ntp_features::kNtpCustomizeChromeAutoShownMaxCount.Get();
+       ++i) {
     OpenNewTabPageInForegroundAndWaitForLoad();
     EXPECT_TRUE(IsCustomizeChromeEntryShowing());
   }
